Crate drone_core

The core crate for Drone, an Embedded Operating System.

Documentation

Drone Book

API documentation

Usage

Add the crate to your Cargo.toml dependencies:

[dependencies]
drone-core = { version = "0.12.0" }

Add or extend std feature as follows:

[features]
std = ["drone-core/std"]

Modules

bitfield
A Bitfield is an integer value treated as a sequence of bits, which can be toggled individually.

ffi
Utilities related to FFI bindings.

fib
The Fibers module.

heap
Dynamic memory allocation.

inventory
A zero-cost abstraction to track various resource states with the type-system.

io
Traits, helpers, and type definitions for core I/O functionality.

log
Debug logging facade.

mem
Basic functions for dealing with memory.

periph
Peripheral is a group of memory-mapped registers or their fields.

prelude
The Drone Prelude.

proc_loop
This module provides interface to wrap a stackful synchronous code into an asynchronous command loop.

reg
The Memory-Mapped Registers module.

sync
Useful synchronization primitives.

thr
The Threads module.

token
The Token trait and its common patterns.

Macros

dbg
Prints and returns the value of a given expression for quick and dirty debugging.

eprint
Prints to the log port #1, if a debug probe is connected.

eprintln
Prints to the log port #1, with a newline, if a debug probe is connected.

heap
Defines dynamic memory structures.

periph
Defines a new generic peripheral.

print
Prints to the log port #0, if the debug probe is connected.

println
Prints to the log port #0, with a newline, if a debug probe is connected.

reg
Defines a memory-mapped register.

thr
Defines the thread type.