Trait drone_core::bitfield::Bits

pub trait Bits where
    Self: Sized + Debug + Copy + PartialOrd + Not<Output = Self> + Sub<Output = Self> + BitOr<Output = Self> + BitXor<Output = Self> + BitAnd<Output = Self> + Shl<Self, Output = Self> + Shr<Self, Output = Self>, {
    fn from_usize(bits: usize) -> Self;

    fn width() -> Self;

    fn is_zero(self) -> bool;
}

An integer interface for Bitfield.

See the module level documentation for details.

Required methods

fn from_usize(bits: usize) -> Self

Creates a new value with the bits of bits.

fn width() -> Self

Returns the width of the integer type in bits.

fn is_zero(self) -> bool

Returns true if all bits of the value are cleared.

Loading content...