Function drone_core::ffi::libc::calloc

#[no_mangle]pub unsafe extern "C" fn calloc(nmemb: size_t, size: size_t) -> *mut c_void

Allocates memory for an array of nmemb elements of size bytes each and returns a pointer to the allocated memory. The memory is set to zero. If nmemb or size is 0, then it returns either NULL, or a unique pointer value that can later be successfully passed to free.

Safety

This function works with raw pointers.