Trait drone_core::fib::ThrFiberClosure

pub trait ThrFiberClosure: ThrToken {
    fn add_fn<F, R>(self, f: F)
    where
        F: FnMut() -> FiberState<(), R>,
        F: Send + 'static,
        R: ReturnNone,
    { ... }

    fn add_once<F>(self, f: F)
    where
        F: FnOnce(),
        F: Unpin + Send + 'static,
    { ... }
}

Extends ThrToken types with add_fn and add_once methods.

Provided methods

fn add_fn<F, R>(self, f: F) where
F: FnMut() -> FiberState<(), R>,
F: Send + 'static,
R: ReturnNone,

Adds a fiber for the closure f to the fiber chain.

fn add_once<F>(self, f: F) where
F: FnOnce(),
F: Unpin + Send + 'static,

Adds a fiber for the closure f to the fiber chain.

Loading content...