Trait drone_core::fib::ThrFiberFuture

pub trait ThrFiberFuture: ThrToken {
    pub fn add_future<F, Y, T>(self, fib: F) -> FiberFuture<T>
    where
        F: Fiber<Input = (), Yield = Y, Return = T>,
        Y: YieldNone,
        F: Send + 'static,
        T: Send + 'static,
    { ... }
}

Extends ThrToken types with add_future method.

Provided methods

pub fn add_future<F, Y, T>(self, fib: F) -> FiberFuture<T> where
F: Fiber<Input = (), Yield = Y, Return = T>,
Y: YieldNone,
F: Send + 'static,
T: Send + 'static,

Adds the fiber fib to the fiber chain and returns a future, which resolves on completion of the fiber.

Loading content...