Trait drone_core::fib::ThrFiberGen

pub trait ThrFiberGen: ThrToken {
    fn add<G>(self, gen: G)
    where
        G: Generator<Yield = (), Return = ()>,
        G: Send + 'static,
    { ... }

    fn add_factory<C, G>(self, factory: C)
    where
        C: FnOnce() -> G + Send + 'static,
        G: Generator<Yield = (), Return = ()>,
        G: 'static,
    { ... }
}

Extends ThrToken types with add and add_factory methods.

Provided methods

fn add<G>(self, gen: G) where
G: Generator<Yield = (), Return = ()>,
G: Send + 'static,

Adds a fiber for the generator gen to the fiber chain.

fn add_factory<C, G>(self, factory: C) where
C: FnOnce() -> G + Send + 'static,
G: Generator<Yield = (), Return = ()>,
G: 'static,

Adds a fiber for the generator returned by factory to the fiber chain.

This method is useful for non-Send fibers.

Loading content...