Trait drone_core::fib::ThrFiberStreamPulse

pub trait ThrFiberStreamPulse: ThrToken {
    fn add_saturating_pulse_stream<F>(self, fib: F) -> FiberStreamPulse
    where
        F: Fiber<Input = (), Yield = Option<usize>, Return = Option<usize>>,
        F: Send + 'static,
    { ... }

    fn add_saturating_pulse_stream_factory<C, F>(
        self, 
        factory: C
    ) -> FiberStreamPulse
    where
        C: FnOnce() -> F + Send + 'static,
        F: Fiber<Input = (), Yield = Option<usize>, Return = Option<usize>>,
        F: 'static,
    { ... }

    fn add_pulse_try_stream<O, F, E>(
        self, 
        overflow: O, 
        fib: F
    ) -> TryFiberStreamPulse<E>
    where
        O: Fn() -> Result<(), E>,
        F: Fiber<Input = (), Yield = Option<usize>, Return = Result<Option<usize>, E>>,
        O: Send + 'static,
        F: Send + 'static,
        E: Send + 'static,
    { ... }

    fn add_pulse_try_stream_factory<C, O, F, E>(
        self, 
        overflow: O, 
        factory: C
    ) -> TryFiberStreamPulse<E>
    where
        C: FnOnce() -> F + Send + 'static,
        O: Fn() -> Result<(), E>,
        F: Fiber<Input = (), Yield = Option<usize>, Return = Result<Option<usize>, E>>,
        O: Send + 'static,
        F: 'static,
        E: Send + 'static,
    { ... }
}

Extends ThrToken types with pulse stream methods.

Provided methods

fn add_saturating_pulse_stream<F>(self, fib: F) -> FiberStreamPulse where
F: Fiber<Input = (), Yield = Option<usize>, Return = Option<usize>>,
F: Send + 'static,

Adds the fiber fib to the fiber chain and returns a stream of pulses yielded from the fiber.

fn add_saturating_pulse_stream_factory<C, F>(
self,
factory: C
) -> FiberStreamPulse where
C: FnOnce() -> F + Send + 'static,
F: Fiber<Input = (), Yield = Option<usize>, Return = Option<usize>>,
F: 'static,

Adds the fiber returned by factory to the fiber chain and returns a stream of pulses yielded from the fiber.

This method is useful for non-Send fibers.

fn add_pulse_try_stream<O, F, E>(
self,
overflow: O,
fib: F
) -> TryFiberStreamPulse<E> where
O: Fn() -> Result<(), E>,
F: Fiber<Input = (), Yield = Option<usize>, Return = Result<Option<usize>, E>>,
O: Send + 'static,
F: Send + 'static,
E: Send + 'static,

Adds the fiber fib to the fiber chain and returns a fallible stream of pulses yielded from the fiber.

fn add_pulse_try_stream_factory<C, O, F, E>(
self,
overflow: O,
factory: C
) -> TryFiberStreamPulse<E> where
C: FnOnce() -> F + Send + 'static,
O: Fn() -> Result<(), E>,
F: Fiber<Input = (), Yield = Option<usize>, Return = Result<Option<usize>, E>>,
O: Send + 'static,
F: 'static,
E: Send + 'static,

Adds the fiber returned by factory to the fiber chain and returns a fallible stream of pulses yielded from the fiber.

This method is useful for non-Send fibers.

Loading content...