Trait drone_core::io::Read

pub trait Read<'sess> {
    type Error;
    fn read(
        &'sess mut self, 
        buf: &'sess mut [u8]
    ) -> Pin<Box<dyn Future<Output = Result<usize, Self::Error>> + Send + 'sess>>;
}

The Read trait allows for reading bytes from a source asynchronously.

Associated Types

type Error

The error type returned by Read::read.

Loading content...
Required methods

fn read(
&'sess mut self,
buf: &'sess mut [u8]
) -> Pin<Box<dyn Future<Output = Result<usize, Self::Error>> + Send + 'sess>>

Pull some bytes from this source into the specified buffer asynchronously, eventually returning how many bytes were read.

Loading content...