Module drone_core::proc_loop

This module provides interface to wrap a stackful synchronous code into an asynchronous command loop.

NOTE A Drone platform crate may re-export this module with its own additions under the same name, in which case it should be used instead.

Enums

Out
Sess::Fiber output.

Traits

Context
A token that allows suspending synchronous code.

ProcLoop
The trait for declaring a synchronous command loop.

Sess
A session type for the synchronous command loop ProcLoop.

Unions

In
Sess::Fiber input.