Module drone_core::reg::prelude

The Memory-Mapped Registers prelude.

The purpose of this module is to alleviate imports of many common reg traits by adding a glob import to the top of reg heavy modules:

use drone_core::reg::prelude::*;

Re-exports

pub use crate::reg::field::RRRegField;

pub use crate::reg::field::RegField;

pub use crate::reg::field::RoRRegField;

pub use crate::reg::field::WWRegField;

pub use crate::reg::field::WoWRegField;

pub use crate::reg::tag::Crt;

pub use crate::reg::tag::RegAtomic;

pub use crate::reg::tag::RegOwned;

pub use crate::reg::tag::RegTag;

pub use crate::reg::tag::Srt;

pub use crate::reg::tag::Urt;

pub use crate::reg::RReg;

pub use crate::reg::Reg;

pub use crate::reg::RegHold;

pub use crate::reg::RoReg;

pub use crate::reg::WReg;

pub use crate::reg::WoReg;

pub use crate::reg::field::RRRegFieldBit as _;

pub use crate::reg::field::RRRegFieldBits as _;

pub use crate::reg::field::RegFieldBit as _;

pub use crate::reg::field::RegFieldBits as _;

pub use crate::reg::field::WWRegFieldBit as _;

pub use crate::reg::field::WWRegFieldBits as _;

pub use crate::reg::field::WoWoRegField as _;

pub use crate::reg::field::WoWoRegFieldBit as _;

pub use crate::reg::field::WoWoRegFieldBits as _;

pub use crate::reg::RegRef as _;

pub use crate::reg::RwRegUnsync as _;

pub use crate::reg::WRegAtomic as _;

pub use crate::reg::WRegUnsync as _;