Trait drone_core::reg::WRegAtomic

pub trait WRegAtomic<'a, T: RegAtomic>: WReg<T> + RegRef<'a, T> {
    fn store<F>(&'a self, f: F)
    where
        F: for<'b> FnOnce(&'b mut Self::Hold) -> &'b mut Self::Hold;

    fn store_val(&self, val: Self::Val);

    fn reset(&'a self);
}

Atomic operations for writable register.

Required methods

fn store<F>(&'a self, f: F) where
F: for<'b> FnOnce(&'b mut Self::Hold) -> &'b mut Self::Hold,

Passes the reset value to the closure f, then writes the result of the closure into the register memory.

See also store_val.

fn store_val(&self, val: Self::Val)

Writes an opaque value val into the register memory.

See also store.

fn reset(&'a self)

Writes the reset value into the register memory.

Loading content...