Trait drone_core::reg::WRegUnsync

pub trait WRegUnsync<'a>: WReg<Urt> + RegRef<'a, Urt> {
    fn store<F>(&'a mut self, f: F)
    where
        F: for<'b> FnOnce(&'b mut Self::Hold) -> &'b mut Self::Hold;

    fn store_reg<F>(&'a mut self, f: F)
    where
        F: for<'b> FnOnce(&'b Self, &'b mut Self::Val);

    fn store_val(&mut self, val: Self::Val);

    fn store_bits(&mut self, bits: <Self::Val as Bitfield>::Bits);

    fn reset(&'a mut self);
}

Non-atomic operations for writable register.

Required methods

fn store<F>(&'a mut self, f: F) where
F: for<'b> FnOnce(&'b mut Self::Hold) -> &'b mut Self::Hold,

Passes the reset value to the closure f, then writes the result of the closure into the register memory.

See also store_reg, store_val, store_bits.

fn store_reg<F>(&'a mut self, f: F) where
F: for<'b> FnOnce(&'b Self, &'b mut Self::Val),

Passes a reference to this register token and the reset value to the closure f, then writes the modified value into the register memory.

See also store, store_val, store_bits.

fn store_val(&mut self, val: Self::Val)

Writes an opaque value val into the register memory.

See also store, store_bits.

fn store_bits(&mut self, bits: <Self::Val as Bitfield>::Bits)

Writes raw bits into the register memory.

See also store, store_val.

fn reset(&'a mut self)

Writes the reset value into the register memory.

Loading content...