Struct drone_core::sync::spsc::ring::Receiver

#[must_use = "futures do nothing unless you `.await` or poll them"]
pub struct Receiver<T, E> { /* fields omitted */ }

The receiving-half of ring::channel.

Methods

impl<T, E> Receiver<T, E>

pub fn close(&mut self)

Gracefully close this receiver, preventing any subsequent attempts to send to it.

Any send operation which happens after this method returns is guaranteed to fail. After calling this method, you can use Receiver::poll to determine whether a message had previously been sent.

pub fn try_recv(&mut self) -> Result<Option<T>, E>

Attempts to receive a value outside of the context of a task.

Does not schedule a task wakeup or have any other side effects.

A return value of Ok(None) must be considered immediately stale (out of date) unless close has been called first.