Module drone_core::thr::prelude

The Threads prelude.

The purpose of this module is to alleviate imports of many common thread token traits by adding a glob import to the top of thread token heavy modules:

use drone_core::thr::prelude::*;

Re-exports

pub use crate::thr::ThrToken;

pub use crate::fib::ThrFiberClosure as _;

pub use crate::fib::ThrFiberFuture as _;

pub use crate::fib::ThrFiberGen as _;

pub use crate::fib::ThrFiberStreamPulse as _;

pub use crate::fib::ThrFiberStreamRing as _;