Trait drone_core::fib::ThrFiberStreamPulse

pub trait ThrFiberStreamPulse: ThrToken {
    fn add_stream_pulse_skip<F>(self, fib: F) -> FiberStreamPulse
    where
        F: Fiber<Input = (), Yield = Option<usize>, Return = Option<usize>>,
        F: Send + 'static,
    { ... }

    fn add_stream_pulse<O, F, E>(
        self, 
        overflow: O, 
        fib: F
    ) -> TryFiberStreamPulse<E>
    where
        O: Fn() -> Result<(), E>,
        F: Fiber<Input = (), Yield = Option<usize>, Return = Result<Option<usize>, E>>,
        O: Send + 'static,
        F: Send + 'static,
        E: Send + 'static,
    { ... }
}

Extends ThrToken types with add_stream_pulse methods.

Provided methods

fn add_stream_pulse_skip<F>(self, fib: F) -> FiberStreamPulse where
F: Fiber<Input = (), Yield = Option<usize>, Return = Option<usize>>,
F: Send + 'static,

Adds the fiber fib to the fiber chain and returns a stream of pulses yielded from the fiber.

fn add_stream_pulse<O, F, E>(
self,
overflow: O,
fib: F
) -> TryFiberStreamPulse<E> where
O: Fn() -> Result<(), E>,
F: Fiber<Input = (), Yield = Option<usize>, Return = Result<Option<usize>, E>>,
O: Send + 'static,
F: Send + 'static,
E: Send + 'static,

Adds the fiber fib to the fiber chain and returns a fallible stream of pulses yielded from the fiber.

Loading content...