Struct drone_core::inventory::Inventory

#[repr(transparent)]
pub struct Inventory<T: Item, C> { /* fields omitted */ }

The inventory wrapper for T. Parameter C encodes the number of emitted tokens.

See the module-level documentation for details.

Methods

impl<T: Item> Inventory<T, Count0>

pub fn new(item: T) -> Self

Creates a new Inventory in the inactive state with zero tokens emitted.

item should contain some form of token.

pub fn free(inventory: Self) -> T

Drops inventory and returns the stored item.

pub fn guard(inventory: &mut Self) -> Guard<T>

Creates an RAII scoped guard.

The item should be already in its active state. The returned guard will call Item::teardown on drop.

pub fn teardown(inventory: &mut Self)

Calls Item::teardown for the stored item.

impl<T: Item, C> Inventory<T, C>

pub fn inventory_token(&self) -> &Token<T>

Returns a reference to Token<T>. While the reference exists, the item is always in its active state.

impl<T: Item> Inventory<T, Count0>

pub fn share1(self) -> (Inventory<T, Count1>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

impl<T: Item> Inventory<T, Count1>

pub fn share1(self) -> (Inventory<T, Count2>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share2(self) -> (Inventory<T, Count3>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share3(self) -> (Inventory<T, Count4>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share4(
self
) -> (Inventory<T, Count5>, Token<T>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share5(
self
) -> (Inventory<T, Count6>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share6(
self
) -> (Inventory<T, Count7>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share7(
self
) -> (Inventory<T, Count8>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share8(
self
) -> (Inventory<T, Count9>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn merge1(self, a: Token<T>) -> Inventory<T, Count0>

Consumes a token and returns a new inventory object with decreased counter in its type.

impl<T: Item> Inventory<T, Count2>

pub fn share1(self) -> (Inventory<T, Count3>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share2(self) -> (Inventory<T, Count4>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share3(self) -> (Inventory<T, Count5>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share4(
self
) -> (Inventory<T, Count6>, Token<T>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share5(
self
) -> (Inventory<T, Count7>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share6(
self
) -> (Inventory<T, Count8>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share7(
self
) -> (Inventory<T, Count9>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn merge1(self, a: Token<T>) -> Inventory<T, Count1>

Consumes a token and returns a new inventory object with decreased counter in its type.

impl<T: Item> Inventory<T, Count3>

pub fn share1(self) -> (Inventory<T, Count4>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share2(self) -> (Inventory<T, Count5>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share3(self) -> (Inventory<T, Count6>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share4(
self
) -> (Inventory<T, Count7>, Token<T>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share5(
self
) -> (Inventory<T, Count8>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share6(
self
) -> (Inventory<T, Count9>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn merge1(self, a: Token<T>) -> Inventory<T, Count2>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge2(self, a: Token<T>, b: Token<T>) -> Inventory<T, Count1>

Consumes a token and returns a new inventory object with decreased counter in its type.

impl<T: Item> Inventory<T, Count4>

pub fn share1(self) -> (Inventory<T, Count5>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share2(self) -> (Inventory<T, Count6>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share3(self) -> (Inventory<T, Count7>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share4(
self
) -> (Inventory<T, Count8>, Token<T>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share5(
self
) -> (Inventory<T, Count9>, Token<T>, Token<T>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn merge1(self, a: Token<T>) -> Inventory<T, Count3>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge2(self, a: Token<T>, b: Token<T>) -> Inventory<T, Count2>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge3(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>
) -> Inventory<T, Count1>

Consumes a token and returns a new inventory object with decreased counter in its type.

impl<T: Item> Inventory<T, Count5>

pub fn share1(self) -> (Inventory<T, Count6>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share2(self) -> (Inventory<T, Count7>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share3(self) -> (Inventory<T, Count8>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share4(
self
) -> (Inventory<T, Count9>, Token<T>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn merge1(self, a: Token<T>) -> Inventory<T, Count4>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge2(self, a: Token<T>, b: Token<T>) -> Inventory<T, Count3>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge3(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>
) -> Inventory<T, Count2>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge4(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>,
d: Token<T>
) -> Inventory<T, Count1>

Consumes a token and returns a new inventory object with decreased counter in its type.

impl<T: Item> Inventory<T, Count6>

pub fn share1(self) -> (Inventory<T, Count7>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share2(self) -> (Inventory<T, Count8>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share3(self) -> (Inventory<T, Count9>, Token<T>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn merge1(self, a: Token<T>) -> Inventory<T, Count5>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge2(self, a: Token<T>, b: Token<T>) -> Inventory<T, Count4>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge3(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>
) -> Inventory<T, Count3>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge4(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>,
d: Token<T>
) -> Inventory<T, Count2>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge5(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>,
d: Token<T>,
e: Token<T>
) -> Inventory<T, Count1>

Consumes a token and returns a new inventory object with decreased counter in its type.

impl<T: Item> Inventory<T, Count7>

pub fn share1(self) -> (Inventory<T, Count8>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn share2(self) -> (Inventory<T, Count9>, Token<T>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn merge1(self, a: Token<T>) -> Inventory<T, Count6>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge2(self, a: Token<T>, b: Token<T>) -> Inventory<T, Count5>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge3(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>
) -> Inventory<T, Count4>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge4(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>,
d: Token<T>
) -> Inventory<T, Count3>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge5(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>,
d: Token<T>,
e: Token<T>
) -> Inventory<T, Count2>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge6(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>,
d: Token<T>,
e: Token<T>,
f: Token<T>
) -> Inventory<T, Count1>

Consumes a token and returns a new inventory object with decreased counter in its type.

impl<T: Item> Inventory<T, Count8>

pub fn share1(self) -> (Inventory<T, Count9>, Token<T>)

Returns a token and a new inventory object with increased counter in its type.

pub fn merge1(self, a: Token<T>) -> Inventory<T, Count7>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge2(self, a: Token<T>, b: Token<T>) -> Inventory<T, Count6>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge3(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>
) -> Inventory<T, Count5>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge4(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>,
d: Token<T>
) -> Inventory<T, Count4>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge5(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>,
d: Token<T>,
e: Token<T>
) -> Inventory<T, Count3>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge6(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>,
d: Token<T>,
e: Token<T>,
f: Token<T>
) -> Inventory<T, Count2>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge7(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>,
d: Token<T>,
e: Token<T>,
f: Token<T>,
g: Token<T>
) -> Inventory<T, Count1>

Consumes a token and returns a new inventory object with decreased counter in its type.

impl<T: Item> Inventory<T, Count9>

pub fn merge1(self, a: Token<T>) -> Inventory<T, Count8>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge2(self, a: Token<T>, b: Token<T>) -> Inventory<T, Count7>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge3(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>
) -> Inventory<T, Count6>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge4(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>,
d: Token<T>
) -> Inventory<T, Count5>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge5(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>,
d: Token<T>,
e: Token<T>
) -> Inventory<T, Count4>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge6(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>,
d: Token<T>,
e: Token<T>,
f: Token<T>
) -> Inventory<T, Count3>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge7(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>,
d: Token<T>,
e: Token<T>,
f: Token<T>,
g: Token<T>
) -> Inventory<T, Count2>

Consumes a token and returns a new inventory object with decreased counter in its type.

pub fn merge8(
self,
a: Token<T>,
b: Token<T>,
c: Token<T>,
d: Token<T>,
e: Token<T>,
f: Token<T>,
g: Token<T>,
h: Token<T>
) -> Inventory<T, Count1>

Consumes a token and returns a new inventory object with decreased counter in its type.